requirements.txt
rich>=13.7.0
matplotlib>=3.8.0
Pillow>=10.1.0
requests>=2.31.0
